tree/refactor - remove unused files
