+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eqworks/lumen-tree",
-  "version": "0.1.0-alpha.9",
+  "version": "0.1.0-alpha.11",
   "description": "Lumen-Tree",
   "main": "dist/index.js",
   "source": "src/index.js",
