+++ src/components/icons/index.js
@@ -1,5 +1,4 @@
 export { default as ChevronDown } from './chevron-down'
 export { default as ChevronUp } from './chevron-up'
 export { default as ChevronRight } from './chevron-right'
-export { default as ChevronDownOutlined } from './chevron-down-outlined'
 export { default as Check } from './check'
