--- src/components/index.js
@@ -1,2 +1,0 @@
-export { default as Dialog } from './dialog'
-export { default as List } from './list'
