+++ src/components/menu-list/index.js
@@ -1,0 +1,2 @@
+export { default as List } from './list'
+export { default as ListItem } from './list-item'
