+++ src/components/tree/tree-menu.js
@@ -2,8 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import { ChevronRight, Check } from '../icons'
-import List from '../list/list'
-import ListItem from '../list/list-item'
+import { List, ListItem } from '../menu-list'
 
 
 const renderMenu = ({ parentLists = [], menuOptions, ...rest }) => {
