+++ src/index.js
@@ -1,1 +1,4 @@
+import './styles/index.css'
+
+
 export { default as Tree } from './components/tree'
